# src/main/resources/openapi-doc.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SaltyVote API",
        "description": "API documentation for the saltyvote application.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "Bearer Authentication": []
        }
    ],
    "paths": {
        "/api/v1/votes/airbnb/{voteId}": {
            "get": {
                "tags": [
                    "vote-creation-controller"
                ],
                "operationId": "getAirbnbVote",
                "parameters": [
                    {
                        "name": "voteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoteWithAirbnbOptionsDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "vote-creation-controller"
                ],
                "operationId": "updateAirbnbVote",
                "parameters": [
                    {
                        "name": "voteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAirbnbVoteDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {}
                    }
                }
            },
            "delete": {
                "tags": [
                    "vote-creation-controller"
                ],
                "operationId": "deleteVote",
                "parameters": [
                    {
                        "name": "voteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {}
                    }
                }
            }
        },
        "/api/v1/votes/{voteId}/collaborators": {
            "get": {
                "tags": [
                    "vote-collaborator-controller"
                ],
                "operationId": "getCollaborators",
                "parameters": [
                    {
                        "name": "voteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CollaboratorDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "vote-collaborator-controller"
                ],
                "operationId": "addCollaborator",
                "parameters": [
                    {
                        "name": "voteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddCollaboratorDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CollaboratorDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/votes/airbnb": {
            "get": {
                "tags": [
                    "vote-creation-controller"
                ],
                "operationId": "getAllVotes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VoteListDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "vote-creation-controller"
                ],
                "operationId": "createNewAirbnbVote",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAirbnbVoteDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int64"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/votes/airbnb/{voteId}/options": {
            "post": {
                "tags": [
                    "vote-creation-controller"
                ],
                "operationId": "addVoteOption",
                "parameters": [
                    {
                        "name": "voteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AirbnbVoteOptionDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int64"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/votes/airbnb/{voteId}/status": {
            "patch": {
                "tags": [
                    "vote-creation-controller"
                ],
                "operationId": "markVoteAsDone",
                "parameters": [
                    {
                        "name": "voteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "done",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {}
                    }
                }
            }
        },
...
        "/api/v1/votes/{voteId}/collaborators/{collaboratorId}": {
            "delete": {
                "tags": [
                    "vote-collaborator-controller"
                ],
                "operationId": "removeCollaborator",
                "parameters": [
                    {
                        "name": "voteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "collaboratorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {}
                    }
                }
            }
        },
        "/api/v1/votes/airbnb/{voteId}/options/{optionId}": {
            "delete": {
                "tags": [
                    "vote-creation-controller"
                ],
                "operationId": "deleteVoteOption",
                "parameters": [
                    {
                        "name": "voteId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "optionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {}
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AirbnbVoteOptionData": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "benefits": {
                        "type": "string"
                    },
                    "disadvantages": {
                        "type": "string"
                    },
                    "travelTime": {
                        "type": "number",
                        "format": "float"
                    },
                    "totalPrice": {
                        "type": "number",
                        "format": "float"
                    },
                    "country": {
                        "type": "string"
                    },
                    "flightNeeded": {
                        "type": "boolean"
                    },
                    "airbnbPrice": {
                        "type": "number",
                        "format": "float"
                    },
                    "airbnbLink": {
                        "type": "string"
                    }
                },
                "required": [
                    "airbnbLink",
                    "airbnbPrice",
                    "benefits",
                    "country",
                    "description",
                    "disadvantages",
                    "flightNeeded",
                    "totalPrice",
                    "travelTime"
                ]
            },
            "AirbnbVoteOptionDto": {
                "type": "object",
                "properties": {
                    "label": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "benefits": {
                        "type": "string"
                    },
                    "disadvantages": {
                        "type": "string"
                    },
                    "travelTime": {
                        "type": "number",
                        "format": "float"
                    },
                    "totalPrice": {
                        "type": "number",
                        "format": "float"
                    },
                    "country": {
                        "type": "string"
                    },
                    "flightNeeded": {
                        "type": "boolean"
                    },
                    "airbnbPrice": {
                        "type": "number",
                        "format": "float"
                    },
                    "airbnbLink": {
                        "type": "string"
                    },
                    "predefinedPosition": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "airbnbData": {
                        "$ref": "#/components/schemas/AirbnbVoteOptionData"
                    }
                },
                "required": [
                    "airbnbData",
                    "airbnbLink",
                    "airbnbPrice",
                    "benefits",
                    "country",
                    "description",
                    "disadvantages",
                    "flightNeeded",
                    "label",
                    "totalPrice",
                    "travelTime"
                ]
            },
            "CreateAirbnbVoteDto": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "voteType": {
                        "type": "string",
                        "enum": [
                            "RANKING",
                            "SIMPLE"
                        ]
                    },
                    "allowMultiVote": {
                        "type": "boolean"
                    },
                    "allowAnonymousVote": {
                        "type": "boolean"
                    },
                    "done": {
                        "type": "boolean"
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AirbnbVoteOptionDto"
                        }
                    }
                },
                "required": [
                    "allowAnonymousVote",
                    "allowMultiVote",
                    "done",
                    "options",
                    "title",
                    "voteType"
                ]
            },
            "AddCollaboratorDto": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "userId"
                ]
            },
            "CollaboratorDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "id",
                    "userId",
                    "username"
                ]
            },
...
            "VoteListDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "title"
                ]
            },
            "AirbnbVoteOptionResponseDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "label": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AirbnbVoteOptionData"
                    },
                    "preDefinedPosition": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "required": [
                    "id",
                    "label"
                ]
            },
            "VoteWithAirbnbOptionsDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "voteType": {
                        "type": "string",
                        "enum": [
                            "RANKING",
                            "SIMPLE"
                        ]
                    },
                    "allowMultiVote": {
                        "type": "boolean"
                    },
                    "allowAnonymousVote": {
                        "type": "boolean"
                    },
                    "done": {
                        "type": "boolean"
                    },
                    "isCreator": {
                        "type": "boolean"
                    },
                    "isCollaborator": {
                        "type": "boolean"
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AirbnbVoteOptionResponseDto"
                        }
                    }
                },
                "required": [
                    "allowAnonymousVote",
                    "allowMultiVote",
                    "done",
                    "id",
                    "isCollaborator",
                    "isCreator",
                    "options",
                    "title",
                    "voteType"
                ]
            },
...
        }
    }
}